Remove several container classes and inline code
