+++ src/main/java/org/mockito/internal/util/collections/Sets.java
@@ -18,10 +18,6 @@ public abstract class Sets {
         return HashCodeAndEqualsSafeSet.of(mocks);
     }
 
-    public static IdentitySet newIdentitySet() {
-        return new IdentitySet();
-    }
-
     public static <T> Set<T> newSet(T ... elements) {
         if (elements == null) {
             throw new IllegalArgumentException("Expected an array of elements (or empty array) but received a null.");
